@@ -1,11 +1,12 @@
 ﻿using System.Text;
 using ExpensesCounter.Web.DI;
+using ExpensesCounter.Web.Options;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
 using Microsoft.IdentityModel.Tokens;
 
 namespace ExpensesCounter.Web
@@ -22,38 +23,49 @@ namespace ExpensesCounter.Web
         // This method gets called by the runtime. Use this method to add services to the container.
         public void ConfigureServices(IServiceCollection services)
         {
-            services.ResolveDependencies();
-            services.AddContext(Configuration["Connections:DefaultConnection"]);
-            
-            services.AddAuthentication(JwtBearerDefaults.AuthenticationScheme)
-                .AddJwtBearer(options =>
-                {
-                    options.RequireHttpsMetadata = false;
-                    options.TokenValidationParameters = new TokenValidationParameters
-                    {
-                        ValidateIssuer = true,
-                        ValidIssuer = Configuration["AuthOptions:Issuer"],
-                        ValidateAudience = true,
-                        ValidAudience = Configuration["AuthOptions:Audience"],
-                        ValidateLifetime = true,
-                        IssuerSigningKey =
-                            new SymmetricSecurityKey(Encoding.ASCII.GetBytes(Configuration["AuthOptions:SecurityKey"])),
-                        ValidateIssuerSigningKey = true
-                    };
-                });
+            var authSection       = Configuration.GetSection("AuthOptions");
+            var connectionSection = Configuration.GetSection("Connections");
 
-            
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_2);
+            var authOptions       = authSection.Get<AuthOptions>();
+            var connectionOptions = connectionSection.Get<DbConnectionOptions>();
+
+            services.Configure<AuthOptions>(authSection);
+            services.Configure<DbConnectionOptions>(connectionSection);
+
+            services.ResolveDependencies();
+
+            services.AddContext(connectionOptions.DefaultConnection);
+
+            services.AddAuthentication(JwtBearerDefaults.AuthenticationScheme)
+                    .AddJwtBearer(options =>
+                     {
+                         options.RequireHttpsMetadata = false;
+                         options.TokenValidationParameters = new TokenValidationParameters
+                         {
+                             ValidateIssuer   = true,
+                             ValidIssuer      = authOptions.Issuer,
+                             ValidateAudience = true,
+                             ValidAudience    = authOptions.Audience,
+                             ValidateLifetime = true,
+                             IssuerSigningKey =
+                                 new SymmetricSecurityKey(Encoding.ASCII.GetBytes(authOptions.SecurityKey)),
+                             ValidateIssuerSigningKey = true
+                         };
+                     });
+
+            services.AddRouting();
+            services.AddControllers();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             if (env.IsDevelopment()) app.UseDeveloperExceptionPage();
 
             app.UseAuthentication();
-            
-            app.UseMvc();
+
+            app.UseRouting();
+            app.UseEndpoints(endpoints => { endpoints.MapControllers(); });
         }
     }
 }
