@@ -2,17 +2,10 @@ namespace ExpensesCounter.Common.Models.Auth
 {
     public class LoginModel
     {
-        public LoginModel(string email, string password, string passwordConfirm)
-        {
-            Email = email;
-            Password = password;
-            PasswordConfirm = passwordConfirm;
-        }
+        public string Email { get; set; }
 
-        public string Email { get; }
-
-        public string Password { get; }
-        public string PasswordConfirm { get; }
+        public string Password { get; set; }
+        public string PasswordConfirm { get; set; }
 
         public bool IsValid =>
             !string.IsNullOrWhiteSpace(Email) &&
