@@ -19,8 +19,8 @@ namespace ExpensesCounter.Web.BLL.Account
             _securityKey = securityKey;
         }
 
-        public string Issuer { private get; set; }
-        public string Audience { private get; set; }
+        public string   Issuer   { private get; set; }
+        public string   Audience { private get; set; }
         public TimeSpan LifeTime { private get; set; }
 
         public string GenerateToken(IEnumerable<Claim> claims)
@@ -28,15 +28,15 @@ namespace ExpensesCounter.Web.BLL.Account
             var now = DateTime.UtcNow;
             var credentials =
                 new SigningCredentials(new SymmetricSecurityKey(Encoding.ASCII.GetBytes(_securityKey)),
-                    SecurityAlgorithms.HmacSha256);
+                                       SecurityAlgorithms.HmacSha256);
 
             var jwt = new JwtSecurityToken(
-                Issuer,
-                Audience,
-                notBefore: now,
-                claims: claims,
-                expires: now.Add(LifeTime),
-                signingCredentials: credentials);
+                                           Issuer,
+                                           Audience,
+                                           notBefore: now,
+                                           claims: claims,
+                                           expires: now.Add(LifeTime),
+                                           signingCredentials: credentials);
 
             var encodedJwt = new JwtSecurityTokenHandler().WriteToken(jwt);
 
