@@ -12,15 +12,15 @@ namespace ExpensesCounter.Web.DAL.Entities
         public DateTime? Birthday { get; set; }
 
         public string FirstName { get; set; }
-        public string LastName { get; set; }
+        public string LastName  { get; set; }
 
         public string PasswordHash { get; set; }
 
         public IEnumerable<ExpensesListUser> ExpensesLists { get; set; }
-        public IEnumerable<RefreshToken> RefreshTokens { get; set; }
+        public IEnumerable<RefreshToken>     RefreshTokens { get; set; }
 
         public DateTimeOffset RegisterDateUtc { get; set; }
-        public DateTimeOffset LastLoginUtc { get; set; }
+        public DateTimeOffset LastLoginUtc    { get; set; }
 
 
         public static void GenerateExceptionIfNull(User user)
