@@ -13,7 +13,7 @@ namespace ExpensesCounter.Web
         public static IWebHostBuilder CreateWebHostBuilder(string[] args)
         {
             return WebHost.CreateDefaultBuilder(args)
-                .UseStartup<Startup>();
+                          .UseStartup<Startup>();
         }
     }
 }
